+++ packages/lumx-react/src/components/flex-box/FlexBox.stories.tsx
@@ -10,6 +10,7 @@ const { meta, ...stories } = setup({
     component: FlexBox,
     components: { Button, Text, IconButton },
     decorators: { withWrapper, withCombinations },
+    argsConverter: (args) => args,
 });
 
 export default {