+++ packages/lumx-vue/src/index.ts
@@ -1,6 +1,7 @@
 export * from '@lumx/core/js/constants';
 export * from '@lumx/core/js/types';
 
+export * from './components/avatar';
 export * from './components/badge';
 export * from './components/button';
 export * from './components/checkbox';